1.金额输入的精度和小数处理 2.汇率输入的精度和小数处理 3.总额的精度修正

+++ HANDMobileExp/Frameworks/LittleMvc/LMRateFieldCell.h
@@ -22,10 +22,9 @@
 
 
 
-@property (nonatomic, assign) float numberValue;
+@property (nonatomic, assign) double numberValue;
 
-
-@property (nonatomic, assign) float lowerLimit;
+@property (nonatomic, assign) double lowerLimit;
 @property (nonatomic, assign) NSUInteger upperLimit;
 @property(nonatomic) UIKeyboardType keyboardType;
 @property (nonatomic, strong) NSNumberFormatter *numberFormatter;
